backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
websockets==15.0.1
pymongo==4.15.0
qdrant-client==1.15.1
google-generativeai==0.8.5
langgraph==0.6.7
python-multipart==0.0.20
redis==5.2.1
scikit-learn==1.5.2
numpy==2.3.3
orjson==3.11.3
